+++ aws-java-sdk-efs/src/main/java/com/amazonaws/services/elasticfilesystem/AmazonElasticFileSystemAsync.java
@@ -1407,8 +1407,9 @@ public interface AmazonElasticFileSystemAsync extends AmazonElasticFileSystem {
      * resource-based policy and can contain multiple policy statements. A file system always has exactly one file
      * system policy, which can be the default policy or an explicit policy set or updated using this API operation.
      * When an explicit policy is set, it overrides the default policy. For more information about the default file
-     * system policy, see <a href="https://docs.aws.amazon.com/efs/latest/ug/res-based-policies-efs.html">Using
-     * Resource-based Policies with EFS</a>.
+     * system policy, see <a
+     * href="https://docs.aws.amazon.com/efs/latest/ug/iam-access-control-nfs-efs.html#default-filesystempolicy">Default
+     * EFS File System Policy</a>.
      * </p>
      * <p>
      * This operation requires permissions for the <code>elasticfilesystem:PutFileSystemPolicy</code> action.
@@ -1428,8 +1429,9 @@ public interface AmazonElasticFileSystemAsync extends AmazonElasticFileSystem {
      * resource-based policy and can contain multiple policy statements. A file system always has exactly one file
      * system policy, which can be the default policy or an explicit policy set or updated using this API operation.
      * When an explicit policy is set, it overrides the default policy. For more information about the default file
-     * system policy, see <a href="https://docs.aws.amazon.com/efs/latest/ug/res-based-policies-efs.html">Using
-     * Resource-based Policies with EFS</a>.
+     * system policy, see <a
+     * href="https://docs.aws.amazon.com/efs/latest/ug/iam-access-control-nfs-efs.html#default-filesystempolicy">Default
+     * EFS File System Policy</a>.
      * </p>
      * <p>
      * This operation requires permissions for the <code>elasticfilesystem:PutFileSystemPolicy</code> action.