+++ aws-java-sdk-redshift/src/main/java/com/amazonaws/services/redshift/model/ScheduledActionType.java
@@ -32,6 +32,18 @@ public class ScheduledActionType implements Serializable, Cloneable {
      * </p>
      */
     private ResizeClusterMessage resizeCluster;
+    /**
+     * <p>
+     * An action that runs a <code>PauseCluster</code> API operation.
+     * </p>
+     */
+    private PauseClusterMessage pauseCluster;
+    /**
+     * <p>
+     * An action that runs a <code>ResumeCluster</code> API operation.
+     * </p>
+     */
+    private ResumeClusterMessage resumeCluster;
 
     /**
      * <p>
@@ -74,6 +86,86 @@ public class ScheduledActionType implements Serializable, Cloneable {
     }
 
     /**
+     * <p>
+     * An action that runs a <code>PauseCluster</code> API operation.
+     * </p>
+     * 
+     * @param pauseCluster
+     *        An action that runs a <code>PauseCluster</code> API operation.
+     */
+
+    public void setPauseCluster(PauseClusterMessage pauseCluster) {
+        this.pauseCluster = pauseCluster;
+    }
+
+    /**
+     * <p>
+     * An action that runs a <code>PauseCluster</code> API operation.
+     * </p>
+     * 
+     * @return An action that runs a <code>PauseCluster</code> API operation.
+     */
+
+    public PauseClusterMessage getPauseCluster() {
+        return this.pauseCluster;
+    }
+
+    /**
+     * <p>
+     * An action that runs a <code>PauseCluster</code> API operation.
+     * </p>
+     * 
+     * @param pauseCluster
+     *        An action that runs a <code>PauseCluster</code> API operation.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     */
+
+    public ScheduledActionType withPauseCluster(PauseClusterMessage pauseCluster) {
+        setPauseCluster(pauseCluster);
+        return this;
+    }
+
+    /**
+     * <p>
+     * An action that runs a <code>ResumeCluster</code> API operation.
+     * </p>
+     * 
+     * @param resumeCluster
+     *        An action that runs a <code>ResumeCluster</code> API operation.
+     */
+
+    public void setResumeCluster(ResumeClusterMessage resumeCluster) {
+        this.resumeCluster = resumeCluster;
+    }
+
+    /**
+     * <p>
+     * An action that runs a <code>ResumeCluster</code> API operation.
+     * </p>
+     * 
+     * @return An action that runs a <code>ResumeCluster</code> API operation.
+     */
+
+    public ResumeClusterMessage getResumeCluster() {
+        return this.resumeCluster;
+    }
+
+    /**
+     * <p>
+     * An action that runs a <code>ResumeCluster</code> API operation.
+     * </p>
+     * 
+     * @param resumeCluster
+     *        An action that runs a <code>ResumeCluster</code> API operation.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     */
+
+    public ScheduledActionType withResumeCluster(ResumeClusterMessage resumeCluster) {
+        setResumeCluster(resumeCluster);
+        return this;
+    }
+
+    /**
      * Returns a string representation of this object. This is useful for testing and debugging. Sensitive data will be
      * redacted from this string using a placeholder value.
      *
@@ -86,7 +178,11 @@ public class ScheduledActionType implements Serializable, Cloneable {
         StringBuilder sb = new StringBuilder();
         sb.append("{");
         if (getResizeCluster() != null)
-            sb.append("ResizeCluster: ").append(getResizeCluster());
+            sb.append("ResizeCluster: ").append(getResizeCluster()).append(",");
+        if (getPauseCluster() != null)
+            sb.append("PauseCluster: ").append(getPauseCluster()).append(",");
+        if (getResumeCluster() != null)
+            sb.append("ResumeCluster: ").append(getResumeCluster());
         sb.append("}");
         return sb.toString();
     }
@@ -105,6 +201,14 @@ public class ScheduledActionType implements Serializable, Cloneable {
             return false;
         if (other.getResizeCluster() != null && other.getResizeCluster().equals(this.getResizeCluster()) == false)
             return false;
+        if (other.getPauseCluster() == null ^ this.getPauseCluster() == null)
+            return false;
+        if (other.getPauseCluster() != null && other.getPauseCluster().equals(this.getPauseCluster()) == false)
+            return false;
+        if (other.getResumeCluster() == null ^ this.getResumeCluster() == null)
+            return false;
+        if (other.getResumeCluster() != null && other.getResumeCluster().equals(this.getResumeCluster()) == false)
+            return false;
         return true;
     }
 
@@ -114,6 +218,8 @@ public class ScheduledActionType implements Serializable, Cloneable {
         int hashCode = 1;
 
         hashCode = prime * hashCode + ((getResizeCluster() == null) ? 0 : getResizeCluster().hashCode());
+        hashCode = prime * hashCode + ((getPauseCluster() == null) ? 0 : getPauseCluster().hashCode());
+        hashCode = prime * hashCode + ((getResumeCluster() == null) ? 0 : getResumeCluster().hashCode());
         return hashCode;
     }